🚚 Rename: 메인 레이아웃 컴포넌트 layout 폴더 하위로 이동
