+++ src/router/Router.jsx
@@ -1,5 +1,5 @@
 import { Routes, Route, BrowserRouter } from 'react-router';
-import MainLayout from '@/components/common/MainLayout';
+import MainLayout from '@/components/common/layout/MainLayout';
 import List from '@/pages/List';
 import Main from '@/pages/Main';
 import Post from '@/pages/Post';
